+++ zaplogger.docker.yml
@@ -1,20 +1,26 @@
-level: "info"
-encoding: "console"
-development: true
-encoderConfig:
-  messageKey: "Msg"
-  levelKey: "Level"
-  timeKey: "Time"
-  nameKey: "Name"
-  callerKey: "Caller"
-  stacktraceKey: "St"
-  levelEncoder: "capital"
-  timeEncoder: "iso8601"
-  durationEncoder: "string"
-  callerEncoder: "short"
-outputPaths:
-  - "stdout"
-  - "./application.log"
-errorOutputPaths:
-  - "stderr"
-  - "./error.log"
+zap_config: 
+  level: "info"
+  encoding: "console"
+  development: true
+  encoderConfig:
+    messageKey: "Msg"
+    levelKey: "Level"
+    timeKey: "Time"
+    nameKey: "Name"
+    callerKey: "Caller"
+    stacktraceKey: "St"
+    levelEncoder: "capital"
+    timeEncoder: "iso8601"
+    durationEncoder: "string"
+    callerEncoder: "short"
+  outputPaths:
+    - "stdout"
+    - "./application.log"
+  errorOutputPaths:
+    - "stderr"
+    - "./error.log"
+
+log_rotate:
+  maxsize: 3
+  maxage: 7
+  maxbackups: 7
